+++ examples/asyncio/wamp/rawsocket/frontend.py
@@ -6,7 +6,8 @@ log = logging.getLogger('frontend')
 
 sys.path = [os.path.join(os.path.dirname(__file__), '../../../..')]+sys.path
 
-from autobahn.asyncio.wamp import ApplicationSession, ApplicationRunnerRawSocket
+from autobahn.asyncio.wamp import ApplicationSession
+from runner import ApplicationRunnerRawSocket
 from autobahn.wamp import ApplicationError
